removed BookATripPage from components; only in pages directory now
