+++ client/src/pages/BookATripPage/BookATripPage.css
@@ -1,3 +1,7 @@
+.BookATripPage {
+    overflow: hidden;
+}
+
 .BookATripPage__banner {
     width: 100vw;
     height: 231px;
@@ -17,3 +21,15 @@
     justify-content: center;
     padding: 50px;
 }
+
+@media only screen and (max-width: 320px){
+    .BookATripPage__header{
+        text-align: center;
+        font-size: 36px;
+    }
+
+    .BookATripPage{
+       width: 100vw;
+    }
+}
+
