+++ client/src/pages/BookATripPage/BookATripPage.stories.js
@@ -1,7 +1,9 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import BookATripPage from './BookATripPage';
+import Banner__pink from "../../components/Banner/Banner__pink.png";
 
 storiesOf('BookATripPage', module)
   .add("Default", () => <BookATripPage
+    background={Banner__pink}
   />)
